refactor: simplify WorkOS auth configuration and update build scripts
- Remove manual WorkOS client initialization from auth example
- MastraAuthWorkos now handles configuration internally
- Remove unused libsql dependency from agent example
- Add server-adapters to build:auth and test:auth scripts
