+++ package.json
@@ -47,7 +47,7 @@
     "build:evals": "pnpm turbo build --filter ./packages/evals",
     "build:clients": "pnpm turbo build --filter \"./client-sdks/*\"",
     "build:schema-compat": "pnpm turbo build --filter ./packages/schema-compat",
-    "build:auth": "pnpm turbo build --filter \"./auth/*\" --filter \"./packages/auth\"",
+    "build:auth": "pnpm turbo build --filter \"./auth/*\" --filter \"./packages/auth\" --filter \"./server-adapters/*\"",
     "build:observability": "pnpm turbo build --filter \"./observability/*\"",
     "dev:playground": "pnpm turbo run dev --filter ./packages/playground-ui --filter ./packages/playground --filter ./client-sdks/react",
     "generate:integration": "pnpx tsx ./integration-generator/index.ts",
@@ -55,7 +55,7 @@
     "test": "vitest run",
     "test:watch": "vitest watch",
     "test:watch:combined-stores": "pnpm --filter \"./stores/*\" test:watch",
-    "test:auth": "pnpm --filter \"./auth/*\" test",
+    "test:auth": "pnpm --filter \"./auth/*\" test && pnpm test:server-adapters",
     "test:cli": "pnpm --filter ./packages/cli test",
     "test:deployer": "pnpm --filter ./packages/deployer test",
     "test:server": "pnpm --filter ./packages/server test",
